requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.0
openpyxl>=3.1.0
pyyaml>=6.0

# LLM and AI dependencies
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
openai>=1.0.0

# Graph and visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Text processing
nltk>=3.8.0
spacy>=3.7.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
click>=8.1.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For advanced NLP features
transformers>=4.30.0
sentence-transformers>=2.2.0